extract common init
